feat: restructure invoice components and update import paths for consistency
