+++ src/components/Invoice/CustomerCombobox.tsx
@@ -1,0 +1,180 @@
+import {
+  Combobox,
+  ComboboxInput,
+  ComboboxButton,
+  ComboboxOptions,
+  ComboboxOption,
+} from "@headlessui/react";
+import { IconChevronDown, IconCheck } from "@tabler/icons-react";
+import { useState, useRef, useEffect } from "react";
+
+interface SelectOption {
+  id: string;
+  name: string;
+}
+
+interface ComboboxProps {
+  name: string;
+  label: string;
+  value: string[];
+  onChange: (value: string[] | null) => void;
+  options: SelectOption[];
+  query: string;
+  setQuery: React.Dispatch<React.SetStateAction<string>>;
+  onLoadMore: () => void;
+  hasMore: boolean;
+  isLoading: boolean;
+}
+
+export const CustomerCombobox: React.FC<ComboboxProps> = ({
+  name,
+  label,
+  value,
+  onChange,
+  options,
+  setQuery,
+  onLoadMore,
+  hasMore,
+  isLoading,
+}) => {
+  const [selectedCustomer, setSelectedCustomer] = useState<SelectOption | null>(
+    value.length > 0 ? { id: "", name: value[0] } : null
+  );
+  const [searchValue, setSearchValue] = useState("");
+  const searchTimeoutRef = useRef<NodeJS.Timeout>();
+
+  const handleSearch = (searchText: string) => {
+    setSearchValue(searchText);
+
+    // Clear existing timeout
+    if (searchTimeoutRef.current) {
+      clearTimeout(searchTimeoutRef.current);
+    }
+
+    // Set new timeout for debouncing
+    searchTimeoutRef.current = setTimeout(() => {
+      // Reset to first page when searching
+      setQuery(searchText);
+    }, 300);
+  };
+
+  const handleCustomerSelection = (customer: SelectOption | null) => {
+    setSelectedCustomer(customer);
+    onChange(customer ? [customer.name] : null);
+    // Clear search value after selection
+    setSearchValue("");
+  };
+
+  // Update selected customer when value changes externally
+  useEffect(() => {
+    if (
+      value.length > 0 &&
+      (!selectedCustomer || selectedCustomer.name !== value[0])
+    ) {
+      setSelectedCustomer({ id: "", name: value[0] });
+    }
+  }, [value]);
+
+  // Cleanup timeout on unmount
+  useEffect(() => {
+    return () => {
+      if (searchTimeoutRef.current) {
+        clearTimeout(searchTimeoutRef.current);
+      }
+    };
+  }, []);
+
+  // Filter options based on search value
+  const filteredOptions =
+    searchValue === ""
+      ? options
+      : options.filter(
+          (option) =>
+            option.name.toLowerCase().includes(searchValue.toLowerCase()) ||
+            option.id.toLowerCase().includes(searchValue.toLowerCase())
+        );
+
+  return (
+    <div className="my-2 space-y-2">
+      <label htmlFor={name} className="text-sm font-medium text-default-700">
+        {label}
+      </label>
+      <Combobox value={selectedCustomer} onChange={handleCustomerSelection}>
+        <div className="relative">
+          <ComboboxInput
+            className="w-full cursor-input rounded-lg border border-default-300 bg-white py-2 pl-4 pr-10 text-left focus:outline-none focus:border-default-500"
+            displayValue={(customer: SelectOption | null) =>
+              customer?.name || ""
+            }
+            onChange={(event) => handleSearch(event.target.value)}
+            placeholder="Search customers..."
+          />
+          <ComboboxButton className="absolute inset-y-0 right-2 flex items-center pr-2">
+            <IconChevronDown
+              className="h-5 w-5 text-default-400"
+              aria-hidden="true"
+            />
+          </ComboboxButton>
+          <ComboboxOptions className="absolute z-20 w-full p-1 mt-1 border bg-white max-h-60 rounded-lg overflow-auto focus:outline-none shadow-lg">
+            {filteredOptions.length === 0 ? (
+              <div className="relative cursor-default select-none py-2 px-4 text-default-700">
+                {isLoading ? "Loading..." : "No customers found."}
+              </div>
+            ) : (
+              <>
+                {filteredOptions.map((customer) => (
+                  <ComboboxOption
+                    key={customer.id}
+                    value={customer}
+                    className={({ active }) =>
+                      `relative cursor-pointer select-none rounded py-2 pl-4 pr-12 ${
+                        active
+                          ? "bg-default-100 text-default-900"
+                          : "text-default-900"
+                      }`
+                    }
+                  >
+                    {({ selected, active }) => (
+                      <>
+                        <span
+                          className={`block truncate ${
+                            selected ? "font-medium" : "font-normal"
+                          }`}
+                        >
+                          {customer.name}
+                          <span className="ml-2 text-default-400">
+                            ({customer.id})
+                          </span>
+                        </span>
+                        {selected && (
+                          <span className="absolute inset-y-0 right-0 flex items-center pr-3">
+                            <IconCheck
+                              className="h-5 w-5 text-default-600"
+                              aria-hidden="true"
+                            />
+                          </span>
+                        )}
+                      </>
+                    )}
+                  </ComboboxOption>
+                ))}
+                {hasMore && (
+                  <button
+                    onClick={(e) => {
+                      e.preventDefault();
+                      onLoadMore();
+                    }}
+                    className="w-full py-2 text-center text-sm rounded text-sky-500 hover:text-sky-600 hover:bg-default-100 focus:outline-none"
+                    disabled={isLoading}
+                  >
+                    {isLoading ? "Loading more..." : "Load More"}
+                  </button>
+                )}
+              </>
+            )}
+          </ComboboxOptions>
+        </div>
+      </Combobox>
+    </div>
+  );
+};
