+++ src/pages/Invoice/InvoiceDetailsPage.tsx
@@ -35,7 +35,7 @@ import {
 } from "../../utils/invoice/dateUtils";
 import TableEditableCell from "../../components/Table/TableEditableCell";
 import { debounce } from "lodash";
-import { CustomerCombobox } from "../../components/Invois/CustomerCombobox";
+import { CustomerCombobox } from "../../components/Invoice/CustomerCombobox";
 import { useProductsCache } from "../../utils/invoice/useProductsCache";
 import { useSalesmanCache } from "../../utils/catalogue/useSalesmanCache";
