+++ src/pages/pagesRoute.tsx
@@ -11,17 +11,17 @@ import React from "react";
 // Import components here
 import SalesByProductsPage from "./Sales/SalesByProductsPage";
 import SalesBySalesmanPage from "./Sales/SalesBySalesmanPage";
-import InvoiceDetailsPage from "./Invois/InvoiceDetailsPage";
+import InvoiceDetailsPage from "./Invoice/InvoiceDetailsPage";
 import CustomerFormPage from "./Catalogue/CustomerFormPage";
 import CustomerAddPage from "./Catalogue/CustomerAddPage";
 import JobCategoryPage from "./Catalogue/JobCategoryPage";
 import StaffFormPage from "./Catalogue/StaffFormPage";
 import CustomerPage from "./Catalogue/CustomerPage";
 import StaffAddPage from "./Catalogue/StaffAddPage";
-import PDFViewerPage from "./Invois/PDFViewerPage";
+import PDFViewerPage from "./Invoice/PDFViewerPage";
 import ProductPage from "./Catalogue/ProductPage";
-import EInvoicePage from "./Invois/EInvoicePage";
-import InvoicePage from "./Invois/InvoicePage";
+import EInvoicePage from "./Invoice/EInvoicePage";
+import InvoicePage from "./Invoice/InvoicePage";
 import BasicPage from "./Catalogue/BasicPage";
 import StaffPage from "./Catalogue/StaffPage";
 import JobPage from "./Catalogue/JobPage";
